+++ default.nix
@@ -21,6 +21,7 @@ let
     rhyolite-common = ./common;
     rhyolite-datastructures = ./datastructures;
     rhyolite-frontend = ./frontend;
+    rhyolite-logging = ./logging;
   };
 
   # srcs used for overrides
@@ -31,6 +32,8 @@ let
     groundhog = repos.groundhog + /groundhog;
     groundhog-postgresql = repos.groundhog + /groundhog-postgresql;
     groundhog-th = repos.groundhog + /groundhog-th;
+    postgresql-simple = repos.postgresql-simple;  # v0.5.4.0 with a fix
+
     # Newer versions than those in reflex-platform
     gargoyle = repos.gargoyle + /gargoyle;
     gargoyle-postgresql = repos.gargoyle + /gargoyle-postgresql;
@@ -53,6 +56,7 @@ let
       bytestring-trie = haskellLib.dontCheck super.bytestring-trie;
       dependent-monoidal-map = haskellLib.doJailbreak super.dependent-monoidal-map;
       gargoyle-postgresql-nix = haskellLib.overrideCabal super.gargoyle-postgresql-nix { librarySystemDepends = [ pkgs.postgresql ]; };
+      postgresql-simple = haskellLib.dontCheck super.postgresql-simple;
       validation = haskellLib.dontCheck super.validation;
 
       dependent-sum-aeson-orphans = self.callHackageDirect {
@@ -101,6 +105,7 @@ in obelisk // {
       rhyolite-common = ./common;
       rhyolite-datastructures = ./datastructures;
       rhyolite-frontend = ./frontend;
+      rhyolite-logging = ./rhyolite-logging;
       rhyolite-test-suite = ./test;
     };
     shells = rec {